+++ app/components/map/settings/quick-settings/QuickSettingsLayers.vue
@@ -109,6 +109,41 @@
                 </ui-toggle>
             </template>
         </ui-columns-display>
+        <ui-columns-display>
+            <template #col1>
+                <ui-toggle
+                    :model-value="!!store.mapSettings.shortAirportView"
+                    @update:modelValue="setUserMapSettings({ shortAirportView: $event })"
+                >
+                    Short airports view
+
+                    <template #description>
+                        Always show airport info as lines
+                    </template>
+                </ui-toggle>
+            </template>
+            <template #col2>
+                <ui-select
+                    :items="[{ text: 'Disabled', value: false }, { text: 'Enabled (when many)', value: true }, { text: 'Enabled (always)', value: 'always' }]"
+                    :model-value="store.mapSettings.aircraftDeclutter ?? false"
+                    @update:modelValue="setUserMapSettings({ aircraftDeclutter: $event as any })"
+                >
+                    <template #label>
+                        Aircraft Declutter
+                    </template>
+                </ui-select>
+            </template>
+        </ui-columns-display>
+        <ui-select
+            v-if="!isMobile"
+            :items="[{ value: 'bottom-left', text: 'Bottom Left' }, { value: 'top-left', text: 'Top Left' }]"
+            :model-value="store.mapSettings.overlaysPositions ?? 'bottom-left'"
+            @update:modelValue="setUserMapSettings({ overlaysPositions: $event as any })"
+        >
+            <template #label>
+                Minified Overlays position
+            </template>
+        </ui-select>
         <ui-notification
             cookie-name="settings-emergency"
             type="info"
@@ -290,6 +325,7 @@ const dataStore = useDataStore();
 
 const resetActive = ref(false);
 const vatglassesActive = isVatGlassesActive;
+const isMobile = useIsMobile();
 
 // For type safety
 const countersOptions: Record<Required<IUserMapSettings['airportsCounters']>['departuresMode'], string> = {