+++ app/composables/render/airports/layers/airport-style.ts
@@ -91,7 +91,7 @@ export function setAirportStyle(layer: VectorLayer) {
                     ? getSelectedColorFromSettings('approachBookings', true) || radarColors.info300Rgb.join(',')
                     : getSelectedColorFromSettings('approach', true) || radarColors.error300Rgb.join(',');
 
-                let mainTransparency = 0.7;
+                const mainTransparency = 0.7;
 
                 if (isDuplicated) {
                     const isUir = properties.atc.some(x => x.duplicatedBy && uirs.some(y => x.duplicatedBy!.startsWith(y)));
@@ -99,11 +99,6 @@ export function setAirportStyle(layer: VectorLayer) {
                     mainColor = isUir
                         ? getSelectedColorFromSettings('uirs', true) || getCurrentThemeRgbColor('purple600').join(',')
                         : getSelectedColorFromSettings('firs', true) || getCurrentThemeRgbColor('green700').join(',');
-
-                    const transparency = getSelectedColorTransparencyFromSettings(isUir ? 'uirs' : 'firs');
-                    if (transparency) {
-                        mainTransparency = transparency;
-                    }
                 }
 
                 if (properties.selected) {
@@ -197,10 +192,10 @@ export function setAirportStyle(layer: VectorLayer) {
                 const width = 14;
                 const offsetX = (properties.index - ((properties.totalCount - 1) / 2)) * (width - 2);
 
-                const styleCacheKey = String(properties.index) + String(letter) + String(properties.facility.booked) + String(properties.totalCount) + String(properties.selected) + String(showAirportDetails);
+                const styleCacheKey = String(properties.index) + String(letter) + String(properties.facility.booked) + String(properties.totalCount) + String(properties.selected) + String(mapStore.compactAirportView);
 
                 if (!styleCache[styleCacheKey]) {
-                    if (!showAirportDetails) {
+                    if (mapStore.compactAirportView) {
                         styleCache[styleCacheKey] = new Style({
                             text: new Text({
                                 font: 'normal 1px/100% Jura, Arial, sans-serif',